test: rename regression tests more expressively
PR-URL: https://github.com/nodejs/node/pull/19495
Refs: https://github.com/nodejs/node/issues/19105
