+++ test/parallel/test-arm-math-illegal-instruction.js
@@ -1,8 +1,11 @@
 'use strict';
+require('../common');
+
+// This test ensures Math functions don't fail with an "illegal instruction"
+// error on ARM devices (primarily on the Raspberry Pi 1)
 // See https://github.com/nodejs/node/issues/1376
 // and https://code.google.com/p/v8/issues/detail?id=4019
 
-require('../common');
 Math.abs(-0.5);
 Math.acos(-0.5);
 Math.acosh(-0.5);
